requirement.txt
asgiref==3.7.2
Django==4.2.2
django-appconf==1.0.5
django-compressor==4.4
django-ranged-response==0.2.0
django-robots==5.0
django-simple-captcha==0.5.17
humanize==4.7.0
mysqlclient==2.2.0
Pillow==9.5.0
python-dotenv==1.0.0
rcssmin==1.1.1
rjsmin==1.2.1
sqlparse==0.4.4
tzdata==2023.3
